requirements.txt
discord.py>=2.6.3
python-dotenv>=1.1.1
aiohttp>=3.12.15
tqdm>=4.66.0